RecurrentMaskable
