+++ docs/index.rst
@@ -35,6 +35,7 @@ RL Baselines3 Zoo also offers a simple interface to train, evaluate agents and d
   modules/crossq
   modules/ppo_mask
   modules/ppo_recurrent
+  modules/ppo_recurrent_mask
   modules/qrdqn
   modules/tqc
   modules/trpo
