+++ sb3_contrib/__init__.py
@@ -4,7 +4,10 @@ from sb3_contrib.ars import ARS
 from sb3_contrib.crossq import CrossQ
 from sb3_contrib.ppo_mask import MaskablePPO
 from sb3_contrib.ppo_recurrent import RecurrentPPO
+from sb3_contrib.ppo_recurrent_mask import RecurrentMaskablePPO
+
 from sb3_contrib.qrdqn import QRDQN
+
 from sb3_contrib.tqc import TQC
 from sb3_contrib.trpo import TRPO
 
@@ -21,4 +24,5 @@ __all__ = [
     "CrossQ",
     "MaskablePPO",
     "RecurrentPPO",
+    "RecurrentMaskablePPO",
 ]
