+++ sb3_contrib/ppo_mask/ppo_mask.py
@@ -283,29 +283,6 @@ class MaskablePPO(OnPolicyAlgorithm):
 
         return True
 
-    def predict(  # type: ignore[override]
-        self,
-        observation: Union[np.ndarray, dict[str, np.ndarray]],
-        state: Optional[tuple[np.ndarray, ...]] = None,
-        episode_start: Optional[np.ndarray] = None,
-        deterministic: bool = False,
-        action_masks: Optional[np.ndarray] = None,
-    ) -> tuple[np.ndarray, Optional[tuple[np.ndarray, ...]]]:
-        """
-        Get the policy action from an observation (and optional hidden state).
-        Includes sugar-coating to handle different observations (e.g. normalizing images).
-
-        :param observation: the input observation
-        :param state: The last hidden states (can be None, used in recurrent policies)
-        :param episode_start: The last masks (can be None, used in recurrent policies)
-            this correspond to beginning of episodes,
-            where the hidden states of the RNN must be reset.
-        :param deterministic: Whether or not to return deterministic actions.
-        :return: the model's action and the next hidden state
-            (used in recurrent policies)
-        """
-        return self.policy.predict(observation, state, episode_start, deterministic, action_masks=action_masks)
-
     def train(self) -> None:
         """
         Update policy using the currently gathered rollout buffer.
