+++ sb3_contrib/ppo_recurrent/__init__.py
@@ -1,4 +1,5 @@
 from sb3_contrib.ppo_recurrent.policies import CnnLstmPolicy, MlpLstmPolicy, MultiInputLstmPolicy
 from sb3_contrib.ppo_recurrent.ppo_recurrent import RecurrentPPO
 
+
 __all__ = ["CnnLstmPolicy", "MlpLstmPolicy", "MultiInputLstmPolicy", "RecurrentPPO"]
