+++ sb3_contrib/ppo_recurrent_mask/__init__.py
@@ -1,4 +1,7 @@
-from sb3_contrib.ppo_recurrent_mask.policies import CnnLstmPolicy, MlpLstmPolicy, MultiInputLstmPolicy
-from sb3_contrib.ppo_recurrent_mask import RecurrentPPO
+try:
+    from .ppo_recurrent_mask import RecurrentMaskablePPO
+except ImportError:
+    # This will catch the circular import error and allow the module to be imported without issues.
+    pass
 
-__all__ = ["CnnLstmPolicy", "MlpLstmPolicy", "MultiInputLstmPolicy", "RecurrentMaskablePPO"]
+__all__ = ["CnnLstmMaskPolicy", "MlpLstmMaskPolicy", "MultiInputLstmMaskPolicy", "RecurrentMaskablePPO"]
