+++ tests/test_cnn.py
@@ -9,7 +9,7 @@ from stable_baselines3.common.envs import FakeImageEnv
 from stable_baselines3.common.utils import zip_strict
 from stable_baselines3.common.vec_env import DummyVecEnv, VecNormalize, VecTransposeImage, is_vecenv_wrapped
 
-from sb3_contrib import QRDQN, TQC, TRPO, MaskablePPO, RecurrentPPO
+from sb3_contrib import QRDQN, TQC, TRPO, MaskablePPO, RecurrentPPO, RecurrentMaskablePPO
 from sb3_contrib.common.wrappers import ActionMasker
 
 
@@ -164,7 +164,7 @@ def test_feature_extractor_target_net(model_class, share_features_extractor):
     params_should_match(original_param, model.critic.parameters())
 
 
-@pytest.mark.parametrize("model_class", [TRPO, MaskablePPO, RecurrentPPO, QRDQN, TQC])
+@pytest.mark.parametrize("model_class", [TRPO, MaskablePPO, RecurrentPPO, RecurrentMaskablePPO, QRDQN, TQC])
 @pytest.mark.parametrize("normalize_images", [True, False])
 def test_image_like_input(model_class, normalize_images):
     """
@@ -181,7 +181,7 @@ def test_image_like_input(model_class, normalize_images):
         channel_first=True,
         discrete=model_class not in {TQC},
     )
-    if model_class == MaskablePPO:
+    if model_class == MaskablePPO or model_class == RecurrentMaskablePPO:
 
         def action_mask_fn(env):
             random_invalid_action = random.randrange(env.action_space.n)
@@ -201,7 +201,7 @@ def test_image_like_input(model_class, normalize_images):
     )
     policy = "CnnLstmPolicy" if model_class == RecurrentPPO else "CnnPolicy"
 
-    if model_class in {TRPO, MaskablePPO, RecurrentPPO}:
+    if model_class in {TRPO, MaskablePPO, RecurrentPPO, RecurrentMaskablePPO}:
         kwargs.update(dict(n_steps=64, batch_size=64))
     else:
         # Avoid memory error when using replay buffer
